Moved 'ParametersAwareInterface' to root namespace alongside 'ParametersTrait'.
